+++ src/external.hpp
@@ -4,6 +4,7 @@
 /*------------------------------------------------------------------------*/
 
 #include "range.hpp"
+#include <unordered_map>
 
 /*------------------------------------------------------------------------*/
 
@@ -96,11 +97,21 @@ struct External {
   void export_learned_unit_clause (int ilit);
   void export_learned_large_clause (const vector<int> &);
 
+  // If there is a listener for fixed assignments.
+
+  FixedAssignmentListener *fixed_listener;
+  
   // If there is an external propagator.
 
   ExternalPropagator *propagator;
 
   vector<bool> is_observed; // Quick flag for each external variable
+
+  // Saved 'forgettable' original clauses coming from the external propagator.
+  // The value of the map starts with a Boolean flag indicating if the clause
+  // is still present or got already deleted, and then followed by the literals
+  // of the clause.
+  unordered_map<uint64_t, vector<int>> forgettable_original;
 
   void add_observed_var (int elit);
   void remove_observed_var (int elit);
@@ -109,6 +120,8 @@ struct External {
   bool observed (int elit);
   bool is_witness (int elit);
   bool is_decision (int elit);
+
+  void force_backtrack (size_t new_level);
 
   //----------------------------------------------------------------------//