+++ src/mobical.cpp
@@ -203,14 +203,14 @@ struct Shared {
 
 /*------------------------------------------------------------------------*/
 
-class MockPropagator : public ExternalPropagator {
+class MockPropagator : public ExternalPropagator, public FixedAssignmentListener {
 private:
   Solver *s = 0;
 
   // MockPropagator parameters
   size_t lemma_per_cb = 2;
   bool logging = false;
-
+  
   struct ExternalLemma {
     size_t id;
     size_t add_count;
@@ -254,7 +254,6 @@ private:
   // Forced lemme addition (falsified lemma in model)
   bool must_add_clause = false;
   size_t must_add_idx;
-
   // Next decision to make
   size_t decision_loc = 0;
 
@@ -346,7 +345,7 @@ public:
       }
       MLOGC ("0" << std::endl);
 
-      add_new_lemma (false);
+      add_new_lemma (true);
       clause.clear ();
     }
   }
@@ -379,7 +378,6 @@ public:
       observed_variables.insert (lit);
 
       s->add_observed_var (lit);
-
       return lit;
     }
     return 0;
@@ -395,13 +393,85 @@ public:
             observed_variables.end ());
   }
 
-  bool compare_trails () { return true; }
+  bool compare_trails () { 
+#ifndef NDEBUG 
+    size_t etrail_inserted = 0;
+
+    std::set<int> etrail = {};  // Trail of the solver
+    std::set<int> efixed = {};  // Fixed assignments in the solver
+
+    size_t otrail_inserted = 0;
+    
+    std::set<int> otrail = {}; // Observed trail
+    std::set<int> ofixed = {}; // Observed fixed assignments
+
+    
+    size_t idx = 0;
+
+    // 1. Collect merged/eliminated variables in case there are:
+    std::vector<int> eq_class = {};
+    // can be an expensive call, avoid if possible
+    bool is_merger = s->internal->get_merged_literals (eq_class);
+    if (is_merger) {
+      for ( const auto& elit: eq_class ) {  
+        if (is_observed_now(elit)) {
+          etrail.insert (elit);
+          etrail_inserted ++;
+        }
+      }
+      idx++; // trail[0] is processed already
+    }
+  
+    // 2. Collect all other variables from trail
+    for (; idx < s->internal->trail.size(); idx++) {
+      int ilit = s->internal->trail[idx];
+      int elit = s->internal->externalize(ilit);
+      if (is_observed_now(elit)) {
+        etrail.insert (elit);
+        etrail_inserted ++;
+      }
+    }
+
+    for (const auto& level : observed_trail) {
+      for (const auto elit : level) {
+        if (is_observed_now(elit)) {
+          // There can be duplicate assignments due to fixed variables
+          // so assert (otrail_inserted == otrail.size()) will not work.
+          assert (otrail.count(elit) == 0 || 
+                  std::find (
+                    observed_fixed.begin (), observed_fixed.end (),
+                    elit) != observed_fixed.end ());
+           
+          otrail.insert (elit);
+          otrail_inserted ++;
+        }
+      }
+    }
+#ifdef LOGGING    
+    if (etrail.size() != otrail.size()) {
+      MLOG ("etrail: ");
+      for (auto const& lit: etrail) MLOGC (lit << " ");
+      MLOGC (std::endl << "otrail: ";);
+      for (auto const& lit: otrail) MLOGC (lit << " ");
+      MLOGC (std::endl);
+    }
+#endif
+    assert (etrail.size() == otrail.size() );
+  
+    assert (etrail == otrail);
+
+#endif
+    return true; 
+  }
   /*-----------------functions for mobical ends ------------------------*/
 
-  /*-------------------------- Observer functions ----------------------*/
+  /*------------------ FixedAssignmentListener functions ---------------------*/
   void notify_fixed_assignment (int lit) {
     MLOG ("notify_fixed_assignment: " << lit << " (current level: "
-                                      << observed_trail.size () - 1 << ")"
+                                      << observed_trail.size () - 1
+                                      << ", current fixed count: "
+                                      << observed_fixed.size()
+                                      << ")"
                                       << std::endl);
 
     assert (std::find (observed_fixed.begin (), observed_fixed.end (),
@@ -414,7 +484,18 @@ public:
       notify_fixed_assignment (lit);
   }
 
-  /* ------------------------ Observer functions end -------------------*/
+  void collect_prev_fixed () {
+#ifndef NDEBUG  
+    MLOG ("collecting previously fixed assignments for the new FixedAssignmentListener: ");
+    
+    std::vector<int> fixed_lits = {};
+    s->internal->get_all_fixed_literals (fixed_lits);
+    MLOGC ("found: " << fixed_lits.size() << " fixed literals" << std::endl);
+    add_prev_fixed(fixed_lits);
+#endif    
+  }
+
+  /* ---------------- FixedAssignmentListener functions end ------------------*/
 
   /* -------------------- ExternalPropagator functions -----------------*/
 
@@ -471,16 +552,16 @@ public:
 
   // Before finalizing the new ipasir-up
   bool cb_has_external_clause () {
-    unsigned red = 0;
-    return cb_has_external_clause (red);
+    bool forgettable = true;
+    return cb_has_external_clause (forgettable);
   }
 
-  bool cb_has_external_clause (unsigned &clause_redundancy) {
+  bool cb_has_external_clause (bool& forgettable) {
     MLOG ("cb_has_external_clause returns: ");
 
     assert (compare_trails ());
 
-    clause_redundancy = 0;
+    forgettable = false;
 
     if (external_lemmas.empty ()) {
       MLOGC ("false (there are no external lemmas)." << std::endl);
@@ -493,11 +574,10 @@ public:
       must_add_clause = false;
       add_lemma_idx = must_add_idx;
 
-      if (external_lemmas[must_add_idx]->forgettable)
-        clause_redundancy = 1;
-
+      forgettable = external_lemmas[must_add_idx]->forgettable;
+        
       MLOGC ("true (forced clause addition, "
-             << "forgettable: " << clause_redundancy
+             << "forgettable: " << forgettable
              << " id: " << add_lemma_idx << ")." << std::endl);
 
       added_lemma_count++;
@@ -520,13 +600,12 @@ public:
       if (!external_lemmas[add_lemma_idx]->add_count &&
           !external_lemmas[add_lemma_idx]->propagation_reason) {
 
-        if (external_lemmas[add_lemma_idx]->forgettable)
-          clause_redundancy = 1;
+        forgettable = external_lemmas[add_lemma_idx]->forgettable;
 
         MLOGC ("true (new lemma was found, "
-               << "forgettable: " << clause_redundancy
-               << " id: " << add_lemma_idx << ")." << std::endl);
-
+            << "forgettable: " << forgettable
+            << " id: " << add_lemma_idx << ")." <<  std::endl);
+        
         added_lemma_count++;
         return true;
       }
@@ -536,7 +615,6 @@ public:
 
       add_lemma_idx++;
     }
-
     MLOGC ("false." << std::endl);
 
     return false;
@@ -573,9 +651,15 @@ public:
         return -1 * new_var;
       }
     }
+
+
     decision_loc++;
 
     if ((decision_loc % observed_variables.size ()) == 0) {
+      if (!(observed_variables.size () % 11)) {
+        MLOG ("cb_decide forces backtracking to level 1" << std::endl);
+        s->force_backtrack (observed_variables.size () % 5);
+      }
       size_t n = decision_loc / observed_variables.size ();
       if (n < observed_variables.size ()) {
         int lit = *std::next (observed_variables.begin (), n);
@@ -657,12 +741,38 @@ public:
       size_t id = add_new_lemma (true);
       external_lemmas[id]->propagation_reason = true;
       reason_map[propagated_lit] = id;
-      clause.clear ();
+      MLOG("new clause added to reason map for " << propagated_lit 
+        << " with id " << id
+        << std::endl);
+      clause.clear();
     }
 
-    MLOG ("cb_propagate returns " << propagated_lit << std::endl);
+    MLOG( "cb_propagate returns " << propagated_lit << std::endl );
 
     return propagated_lit;
+  }
+
+  std::set<int> current_observed_satisfied_set (size_t& lit_sum, int& lowest_lit, int& highest_lit) {
+    
+    lit_sum = 0;
+    lowest_lit = 0;
+    highest_lit = 0;
+    std::set<int> satisfied_literals;
+    
+    for (auto level_lits : observed_trail) {
+      for (auto lit : level_lits) {
+        if (!s->observed (lit))
+          continue;
+
+        satisfied_literals.insert (lit);
+        lit_sum += abs (lit);
+
+        if (!lowest_lit) lowest_lit = lit;
+        highest_lit = lit;
+      }
+    }
+
+    return satisfied_literals;
   }
 
   int cb_add_reason_clause_lit (int plit) {
@@ -684,20 +794,16 @@ public:
     return lit;
   }
 
-  void notify_assignment (int lit, bool is_fixed) {
-    MLOG ("notify assignment: "
-          << lit << " (current level: " << observed_trail.size () - 1
-          << ", is_fixed: " << is_fixed << ")" << std::endl);
-    if (is_fixed) {
-      observed_trail.front ().push_back (lit);
-    } else {
+  void notify_assignment (const std::vector<int>& lits) override { 
+    MLOG ("notified " << lits.size() << " new assignments." << std::endl);
+    for (const auto& lit: lits) {
       observed_trail.back ().push_back (lit);
     }
   }
 
   void notify_new_decision_level () {
-    MLOG ("notify new decision level " << observed_trail.size () - 1
-                                       << " -> " << observed_trail.size ()
+    MLOG ("notify new decision level " << observed_trail.size () -1 << " -> "
+                                       << observed_trail.size ()
                                        << std::endl);
     observed_trail.push_back (std::vector<int> ());
   }
@@ -705,53 +811,36 @@ public:
   void notify_backtrack (size_t new_level) {
     MLOG ("notify backtrack: " << observed_trail.size () - 1 << " -> "
                                << new_level << std::endl);
+    assert (observed_trail.size () > 1 || !new_level);                
     assert (observed_trail.size () == 1 ||
             observed_trail.size () >= new_level + 1);
     while (observed_trail.size () > new_level + 1) {
-      // Remove reason clause of backtracked assignments (keep it as lemma)
-      for (auto lit : observed_trail.back ()) {
-        if (reason_map.find (lit) != reason_map.end ()) {
-          size_t reason_id = reason_map[lit];
-          assert (reason_id < external_lemmas.size ());
-          external_lemmas[reason_id]->propagation_reason = false;
-          external_lemmas[reason_id]->forgettable = true;
-          reason_map.erase (lit);
+      // We can not remove reason clauses of backtracked assignments because
+      // ILB might re-introduces them on the trail. But upon restarts it should
+      // be always safe to clean up the reason database.
+      if (!new_level) {
+        for (auto lit : observed_trail.back ()) {
+          if (reason_map.find (lit) != reason_map.end ()) {
+            size_t reason_id = reason_map[lit];
+            assert (reason_id < external_lemmas.size ());
+            external_lemmas[reason_id]->propagation_reason = false;
+            external_lemmas[reason_id]->forgettable = true;
+            reason_map.erase (lit);
+          }
         }
       }
+#ifndef NDEBUG
+      MLOG("unassign during backtrack from level : " 
+        << observed_trail.size() - 1 );
+      for (auto lit: observed_trail.back()) MLOGC(lit << " ");
+      MLOGC(std::endl);
+#endif
       observed_trail.pop_back ();
     }
   }
 
   /* ----------------- ExternalPropagator functions end ------------------*/
 
-  /* -------------------------- Helper functions ---------------------- */
-  std::set<int> current_observed_satisfied_set (size_t &lit_sum,
-                                                int &lowest_lit,
-                                                int &highest_lit) {
-
-    lit_sum = 0;
-    lowest_lit = 0;
-    highest_lit = 0;
-    std::set<int> satisfied_literals;
-
-    for (auto level_lits : observed_trail) {
-      for (auto lit : level_lits) {
-        if (!s->observed (lit))
-          continue;
-
-        satisfied_literals.insert (lit);
-        lit_sum += abs (lit);
-
-        if (!lowest_lit)
-          lowest_lit = lit;
-        highest_lit = lit;
-      }
-    }
-
-    return satisfied_literals;
-  }
-
-  /* ------------------------ Helper functions end -------------------- */
 };
 
 // This is the class for the Mobical application.
@@ -1241,12 +1330,25 @@ struct ConnectCall : public Call {
     MockPropagator *prev_pointer = 0;
     if (mobical.mock_pointer)
       prev_pointer = mobical.mock_pointer;
-
+#ifdef LOGGING
+    mobical.mock_pointer = new MockPropagator (s, mobical.add_set_log_to_true);
+#else
     mobical.mock_pointer = new MockPropagator (s);
+#endif
     s->connect_external_propagator (mobical.mock_pointer);
+    s->connect_fixed_listener(mobical.mock_pointer);
 
-    if (prev_pointer)
+    if (prev_pointer) {
+      mobical.mock_pointer->add_prev_fixed (prev_pointer->observed_fixed);
       delete prev_pointer;
+    } else {
+      // FixedAssignmentListener does not replay previous fixed assignment,
+      // collect them here explicitly -- EXPENSIVE
+      // In practice FixedAssignmentListener is there from the beginning if 
+      // needed, in mobical we do not want to wire in this.
+      
+      mobical.mock_pointer->collect_prev_fixed ();
+    }
   }
   void print (ostream &o) { o << "connect mock-propagator" << endl; }
   Call *copy () { return new ConnectCall (); }
@@ -1282,20 +1384,6 @@ struct LemmaCall : public Call {
   const char *keyword () { return "lemma"; }
 };
 
-// struct ContinueCall : public Call {
-//   ContinueCall () : Call (CONTINUE) {}
-//   void execute (Solver *&s) {
-//     MockPropagator *mp =
-//         static_cast<MockPropagator *> (s->get_propagator ());
-
-//     if (mp) // || mobical.donot.enforce
-//       mp->push_continue ();
-//   }
-//   void print (ostream &o) { o << "continue" << endl; }
-//   Call *copy () { return new ContinueCall (); }
-//   const char *keyword () { return "continue"; }
-// };
-
 struct DisconnectCall : public Call {
   DisconnectCall () : Call (DISCONNECT) {}
   void execute (Solver *&s) {
@@ -1303,6 +1391,7 @@ struct DisconnectCall : public Call {
         static_cast<MockPropagator *> (s->get_propagator ());
     if (mp)
       mp->remove_new_observed_var ();
+    s->disconnect_fixed_listener ();
     s->disconnect_external_propagator ();
     if (mp) {
       delete mp;
@@ -2106,13 +2195,13 @@ void Trace::generate_propagator (Random &random, int minvars, int maxvars) {
 void Trace::generate_lemmas (Random &random) {
   if (!observed_vars.size ())
     return;
-  int nof_user_propagation_phases = random.pick_int (3, 7);
+  int nof_user_propagation_phases = random.pick_int (4, 7);
 
   for (int p = 0; p < nof_user_propagation_phases; p++) {
     if (random.generate_double () < 0.05) {
       // push_back (new ContinueCall ());
     } else {
-      const int nof_lemmas = random.pick_int (4, 11);
+      const int nof_lemmas = random.pick_int (5, 11);
       const int ovars = observed_vars.size ();
       for (int i = 0; i < nof_lemmas; i++) {
         // Tiny tiny chance to generate an empty lemma
@@ -2417,6 +2506,11 @@ void Trace::generate (uint64_t i, uint64_t s) {
 
     int clauses = range * ratio;
 
+    // TODO: Test empty clause database by uncommenting here
+    // Note that it can lead to unvalid mobical states in the reduced
+    // trace, so always check the original bug-trace too.
+    //if (random.generate_double () < 0.01) clauses = 0;
+    
     minvars = random.pick_int (1, maxvars + 1);
     maxvars = minvars + range;