+++ src/solver.cpp
@@ -313,6 +313,10 @@ static bool tracing_api_calls_through_environment_variable_method;
 #endif
 /*------------------------------------------------------------------------*/
 
+
+static bool tracing_nb_lidrup_env_var_method = false;
+
+
 Solver::Solver () {
 
 #ifndef NTRACING
@@ -347,6 +351,25 @@ Solver::Solver () {
   if (tracing_api_calls_through_environment_variable_method)
     message ("tracing API calls to '%s'", path);
 #endif
+
+  const char *lidrup_path = getenv ("CADICAL_LIDRUP_TRACE");
+  if (!lidrup_path)
+    lidrup_path = getenv ("CADICALLIDRUPTRACE");
+  if (lidrup_path) {
+
+    // if (tracing_nb_lidrup_env_var_method)
+    // FATAL ("can not trace LIDRUP of two solver instances "
+    //   "using environment variable 'CADICAL_LIDRUP_TRACE'");
+    // Here we use the solver interface to setup non-binary IDRUP tracing to
+    // the defined file. Options set by the user can and will overwrite
+    // these settings if neeed be.
+    set ("lidrup", 1);
+    set ("binary", 0);
+    trace_proof (lidrup_path);
+    tracing_nb_lidrup_env_var_method = true;
+  } else {
+    tracing_nb_lidrup_env_var_method = false;
+  }
 }
 
 Solver::~Solver () {
@@ -354,6 +377,8 @@ Solver::~Solver () {
   TRACE ("reset");
   REQUIRE_VALID_OR_SOLVING_STATE ();
   STATE (DELETING);
+
+  tracing_nb_lidrup_env_var_method = false;
 
 #ifdef LOGGING
   //
@@ -711,6 +736,7 @@ int Solver::solve () {
   REQUIRE_READY_STATE ();
   const int res = call_external_solve_and_check_results (false);
   LOG_API_CALL_RETURNS ("solve", res);
+  if (tracing_nb_lidrup_env_var_method) flush_proof_trace(true); 
   return res;
 }
 
@@ -883,13 +909,45 @@ void Solver::disconnect_learner () {
 
 /*===== IPASIR END =======================================================*/
 
+void Solver::connect_fixed_listener (FixedAssignmentListener *fixed_listener) {
+  LOG_API_CALL_BEGIN ("connect_fixed_listener");
+  REQUIRE_VALID_STATE ();
+  REQUIRE (fixed_listener, "can not connect zero fixed listener");
+
+#ifdef LOGGING
+  if (external->fixed_listener)
+    LOG ("connecting new listener of fixed assignments (disconnecting previous one)");
+  else
+    LOG ("connecting new listener of fixed assigments (no previous one)");
+#endif  
+  if (external->fixed_listener)
+    disconnect_fixed_listener ();
+  external->fixed_listener = fixed_listener;
+  // Listeners are treated as real-time listeners, thus previously found fixed
+  // assignments are not sent out (would be rather expensive to recover it
+  // retrospect, see external_propagate.cpp/get_fixed_literals () function).
+  LOG_API_CALL_END ("connect_fixed_listener");
+}
+
+void Solver::disconnect_fixed_listener () {
+  LOG_API_CALL_BEGIN ("disconnect_fixed_listener");
+  REQUIRE_VALID_STATE ();
+#ifdef LOGGING
+  if (external->fixed_listener)
+    LOG ("disconnecting previous listener of fixed assignments");
+  else
+    LOG ("ignoring to disconnect listener of fixed assignments (no previous one)");
+#endif
+  external->fixed_listener = 0;
+  LOG_API_CALL_END ("disconnect_fixed_listener");
+}
+
 /*===== IPASIR-UP BEGIN ==================================================*/
 
 void Solver::connect_external_propagator (ExternalPropagator *propagator) {
   LOG_API_CALL_BEGIN ("connect_external_propagator");
   REQUIRE_VALID_STATE ();
   REQUIRE (propagator, "can not connect zero propagator");
-
 #ifdef LOGGING
   if (external->propagator)
     LOG ("connecting new external propagator (disconnecting previous one)");
@@ -1367,6 +1425,12 @@ bool Solver::is_decision (int lit) {
   bool res = external->is_decision (lit);
   LOG_API_CALL_RETURNS ("is_decision", lit, res);
   return res;
+}
+
+void Solver::force_backtrack (size_t new_level) {
+  TRACE ("force_backtrack", new_level);
+  REQUIRE_VALID_OR_SOLVING_STATE ();
+  external->force_backtrack (new_level);
 }
 
 /*------------------------------------------------------------------------*/